Try synchronization with an AutoResetEvent
Still fragile, sometimes passes now with FireFox. An attempt to
synchronize the test execution with an AutoResetEvent is envisaged,
but not yet reliable.
Uses now a static TestFoxus class to reference not the Main, but the
Blazor component in focus. Usage of the server generated Id for
Selenium (as with WebForms) seems to be promising.
Unit tests with the EF Core DBContext in the VS test runner are broken
for now due to "dependency dll hell" in .NET Core.